Console now runs inside IVLE (without requiring an IFRAME). The separate console server is still spawned, but the client never directly communicates with it.
apps/console: Writes the HTML code for the actual console (the 2 text boxes),
    instead of just a contained div for the IFRAME.

media/console/console.js:
    Removed code to create IFRAME and link to console server.
    Incorporated Tom's console.js code from trunk/console/console.js.
    Modified this code to:
        a) Use functions from util.js, such as make_query_string and
            ajax_call.
        b) Talk to consoleservice/chat instead of the console server.
        c) Pass host and port as arguments (needed by IVLE to talk to the
            console server).
    Also had the "startup" code save the host, port and magic, needed for
    communication with the console server.

media/console/console.css: Added CSS file (derived from Tom's inline CSS in
    the original index.html).

apps/consoleservice: Fixed bugs which caused the wrong text to be sent, and
    also avoid throwing an error for empty text.

Note that trunk/console will still serve the HTML, JavaScript and CSS files if
needed. The next step is to make this just a service and not actually a web
server as well.