Converted Console from an "app" into a "plugin". It can now be plugged in to any app. Added "plugins" directory in www. Added "console" plugin. This contains all of the functionality of what was previously the console app, but modularized so it can be imported by another app.
apps/console: Removed most of the logic (moved to plugins/console). Replaced
with a simple import of the console plugin. Should behave exactly the same.
apps/tutorial: As proof of concept, imported the console plugin. It now
appears at the bottom of the page (yet to make it have "pop up" behaviour).