Console: Configured console to display properly as a "floating" window in the tutorial page. The console now has two "modes" (normal mode and "windowpane" mode). This is switched by the JavaScript, and enables different CSS styling (and will further enable different behaviour).
Styled the console (black). In windowpane mode, there is an additional div at
the bottom which fills up the space, and the console is floated on top of it.
(With position:fixed). This lets it appear at the bottom of the screen no
matter where the user scrolls.

Added console_app.css and console_app.js, which are necessary for things
specific to the "Console" application but not all instances of the console.

Changed the styling and layout of tutorial HTML/CSS to accomodate the console
(it needed to go outside the padded box).