setup.py: Gutted out the config options code. It was getting so there were about 6 points of control on the options, and adding a new option required you to write code in all of those places.
Now there is a single data structure storing all of the config options (their
names, default values, and prompts). The code in the rest of setup.py just
acts on this data structure.

Adding new options now only requires appending a new value to config_options
at the top of the script.