Groups: Added userservice/assign_group call. This allows a user with the CAP_MANAGEGROUPS permission to assign a user into a group. This records the assignment in the database and then regenerates a svn-groups.conf file which specifies which users can access which repository.
This commit introduces the /groups repository (which means a groups directory 
will have to be created in repositories area and the apache config will have to 
be updated to allow the new dav repositories to work - see the changes to 
ivle-svn.conf/ivle-both.conf and install_proc for exact details). There is also 
a new parameter svn_group_conf in the config file so you need to run `setup.py 
config` too.