Expose jQuery in our URL space.
Adds the concept of external media files, which are static files that we
pull in from third-party dependencies and serve under /+media/+external.

A new config option (media/externals/jquery) has been added to specify
where to find jquery.js - it defaults to /usr/share/javascript/jquery,
which is both a generally sane location and also the Debian/Ubuntu one.